+++ Assets/Common/Plugins/Obvious/Soap/Examples/Content/Scripts/ScriptableListPlayer.cs
@@ -2,7 +2,6 @@
 
 namespace Obvious.Soap.Example
 {
-   
     [CreateAssetMenu(menuName = "Soap/Examples/ScriptableLists/Player")]
     public class ScriptableListPlayer : ScriptableList<Player>
     {
@@ -12,5 +11,10 @@ namespace Obvious.Soap.Example
             var playerToDestroy = _list[0];
             Destroy(playerToDestroy.gameObject);
         }
+    }
+    
+    [System.Serializable]
+    public class ScriptableListPlayerReadOnly : ScriptableListReadOnly<ScriptableListPlayer, Player>
+    {
     }
 }