+++ Assets/Project/Scripts/UI_Scripts/Custom_VisualElements/bound-progress-element/threshold-meter/ThresholdMeterElement.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Game.UI.Library {
+    public class ThresholdMeterElement : BoundProgressElementBase{
+
+        protected override string controlUssClassName { get; }
+        protected override void OnProgressChanged(float progressPercent) {
+            
+        }
+    }
+
+}